+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/results/tracing/flat/Action.java
@@ -17,7 +17,6 @@ package org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.tracing.flat;
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.processor.TransactionTrace;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.processor.privateProcessor.PrivateTransactionTrace;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.Quantity;
 import org.hyperledger.besu.ethereum.debug.TraceFrame;
 
@@ -181,22 +180,6 @@ public class Action {
           new Builder()
               .from(trace.getTransaction().getSender().toHexString())
               .value(Quantity.create(trace.getTransaction().getValue()));
-      if (!trace.getTraceFrames().isEmpty()) {
-        final TraceFrame traceFrame = trace.getTraceFrames().get(0);
-        builder.gas(
-            "0x"
-                + Long.toHexString(
-                    traceFrame.getGasRemaining()
-                        + (traceFrame.getPrecompiledGasCost().orElse(0L))));
-      }
-      return builder;
-    }
-
-    public static Builder from(final PrivateTransactionTrace trace) {
-      final Builder builder =
-          new Builder()
-              .from(trace.getPrivateTransaction().getSender().toHexString())
-              .value(Quantity.create(trace.getPrivateTransaction().getValue()));
       if (!trace.getTraceFrames().isEmpty()) {
         final TraceFrame traceFrame = trace.getTraceFrames().get(0);
         builder.gas(
