+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/methods/JsonRpcMethodsFactory.java
@@ -138,8 +138,7 @@ public class JsonRpcMethodsFactory {
                   protocolSchedule,
                   transactionPool,
                   privacyParameters,
-                  filterManager,
-                  protocolContext),
+                  filterManager),
               new PrivxJsonRpcMethods(
                   blockchainQueries, protocolSchedule, transactionPool, privacyParameters),
               new Web3JsonRpcMethods(clientVersion),
