+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/methods/PrivJsonRpcMethodsTest.java
@@ -21,7 +21,6 @@ import static org.hyperledger.besu.ethereum.api.jsonrpc.RpcMethod.PRIV_FIND_PRIV
 import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.when;
 
-import org.hyperledger.besu.ethereum.ProtocolContext;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.filter.FilterManager;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.JsonRpcMethod;
 import org.hyperledger.besu.ethereum.api.query.BlockchainQueries;
@@ -45,20 +44,13 @@ public class PrivJsonRpcMethodsTest {
   @Mock private TransactionPool transactionPool;
   @Mock private PrivacyParameters privacyParameters;
   @Mock private FilterManager filterManager;
-  @Mock private ProtocolContext protocolContext;
-
   private PrivJsonRpcMethods privJsonRpcMethods;
 
   @BeforeEach
   public void setup() {
     privJsonRpcMethods =
         new PrivJsonRpcMethods(
-            blockchainQueries,
-            protocolSchedule,
-            transactionPool,
-            privacyParameters,
-            filterManager,
-            protocolContext);
+            blockchainQueries, protocolSchedule, transactionPool, privacyParameters, filterManager);
 
     lenient().when(privacyParameters.isEnabled()).thenReturn(true);
   }
